Organize 'Go To Adjacent Member' unit tests and share code between C# and VB
