+++ src/EditorFeatures/VisualBasicTest/BasicEditorServicesTest.vbproj
@@ -272,7 +272,7 @@
     <Compile Include="Formatting\Indentation\SmartTokenFormatter_FormatTokenTests.vb" />
     <Compile Include="Formatting\XmlLiterals.Designer.vb" />
     <Compile Include="Outlining\AbstractVisualBasicSyntaxTriviaOutlinerTests.vb" />
-    <Compile Include="VisualBasicGoToAdjacentMemberTests.vb" />
+    <Compile Include="GoToAdjacentMember\VisualBasicGoToAdjacentMemberTests.vb" />
     <Compile Include="KeywordHighlighting\AbstractVisualBasicKeywordHighlighterTests.vb" />
     <Compile Include="KeywordHighlighting\AccessorDeclarationHighlighterTests.vb" />
     <Compile Include="KeywordHighlighting\ConditionalPreprocessorHighlighterTests.vb" />
